requirements.txt
amqp==2.2.2
asn1crypto==0.23.0
billiard==3.5.0.3
celery==4.1.0
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
cryptography==1.7.2
Django==1.11.6
django-celery-results==1.0.1
django-cors-headers==2.1.0
djangorestframework==3.7.1
Faker==0.8.6
idna==2.6
kombu==4.1.0
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
requirements==0.1
simplejson==3.12.0
six==1.11.0
StringGenerator==0.2.0
text-unidecode==1.0
typing==3.5.3.0
urllib3==1.22
vine==1.1.4
gunicorn
